[Chore #158719598] Refactor Auth Code
