+++ index.html
@@ -17,8 +17,8 @@
                <li id="logo"><a href="index.html">Maintenance</a></li>
                            
                <li><a href="index.html">Home</a></li>
-               <li><a href="request-form.html" id="reqForm">Request Service</a></li>
-               <li id="login"><a href="login.html">Login/Signup</a></li> 
+               <li><a href="public/pages/request-form.html" id="reqForm">Request Service</a></li>
+               <li id="login"><a href="public/auth/login.html">Login/Signup</a></li> 
             </ul>
         </div> <!-- navigation menu ends -->
 
@@ -32,8 +32,8 @@
                         <p id="flash"></p>
                     <h1 id="welcome-title">Welcome To Maintenance Tracker App</h2>
                     <p>We offer Maintenance and Repair services . Create account and request for a service.</p> 
-                    <a href="signup.html"><button class="btn-primary" >CREATE ACCOUNT</button></a>
-                    <a href="request-form.html"><button class="btn-success" >REQUEST SERVICE</button></a>  
+                    <a href="public/auth/signup.html"><button class="btn-primary" >CREATE ACCOUNT</button></a>
+                    <a href="public/pages/request-form.html"><button class="btn-success" >REQUEST SERVICE</button></a>  
                 </div>
             </div>
         </div>
