+++ package.json
@@ -4,8 +4,8 @@
   "description": "Build with vannilla js and html5",
   "main": "index.js",
   "scripts": {
-    "build": "webpack --config prod.webpack.config.js -p",
-    "start": "webpack-dev-server",
+    "build": "webpack --config webpack.config.js",
+    "start": "webpack --watch & webpack-dev-server",
     "test": "webpack"
   },
   "pre-commit": [
@@ -35,6 +35,7 @@
     "eslint-loader": "^2.0.0",
     "pre-commit": "^1.2.2",
     "webpack": "^4.13.0",
+    "webpack-cli": "^3.0.8",
     "webpack-dev-server": "^3.1.4"
   }
 }
