+++ public/auth/login.html
@@ -3,8 +3,8 @@
     <head>
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Maintenance Tracker App</title>
-        <link rel="stylesheet" href="assets/css/styles.css">
-        <script src="assets/js/controller.js"></script>
+        <link rel="stylesheet" href="../../assets/css/styles.css">
+        <script src="../../dist/auth.min.js"></script>
         
     </head>
     <body>
@@ -16,8 +16,8 @@
                 
                <li id="logo"><a href="index.html">Maintenance</a></li>
                            
-               <li><a href="index.html">Home</a></li>
-               <li><a href="request-form.html">Request Service </a></li>
+               <li><a href="../../index.html">Home</a></li>
+               <li><a href="../pages/request-form.html">Request Service </a></li>
                <li id="login"><a href="login.html">Login/Signup</a></li>  
             </ul>
         </div> <!-- navigation menu ends -->
