+++ public/pages/request-form.html
@@ -93,7 +93,7 @@
                         if(data.message) {
                             
                             localStorage.setItem("msg","Login To Continue");
-                            window.location.href = 'login.html'
+                            window.location.href = '/public/auth/login.html'
                         }
                     })
